requirements.txt
flask
gunicorn
line-bot-sdk
python-dotenv
SQLAlchemy
requests
openai
langchain
